refactor: avoid '.' imports
