+++ src/builder/adapter/column.ts
@@ -19,7 +19,7 @@ import type {
   IBuilderAdapterHierarchyColumnDescProps,
   IBuilderAdapterNumberColumnDescProps,
   IBuilderAdapterStringColumnDescProps,
-} from '.';
+} from './interfaces';
 
 export function build<T extends IBuilderAdapterColumnDescProps>(props: T, _data?: any[]): IColumnDesc {
   const { column } = props;
