+++ src/model/MappingFunction.ts
@@ -1,8 +1,7 @@
 import { scaleLinear, scaleLog, scalePow, scaleSqrt } from 'd3-scale';
 import { similar } from '../internal';
-import type { IMappingFunction, IMapAbleDesc } from '.';
 import type { ITypeFactory, ITypedDump } from './interfaces';
-import type { IMappingFunctionConstructor } from './INumberColumn';
+import type { IMappingFunction, IMapAbleDesc, IMappingFunctionConstructor } from './INumberColumn';
 
 /**
  * interface of a d3 scale
