+++ src/renderer/utils.ts
@@ -2,8 +2,8 @@ import { MIN_LABEL_WIDTH } from '../constants';
 import { Column, IArrayColumn, IDataRow, ICategoricalLikeColumn, isMapAbleColumn, DEFAULT_COLOR } from '../model';
 import { hsl } from 'd3-color';
 import { cssClass } from '../styles';
-import type { IRenderContext } from '.';
 import type { ISequence } from '../internal';
+import type { IRenderContext } from './interfaces';
 
 /** @internal */
 export function noop() {
